Odstranění matoucích redundantních částí kódu a dat
