+++ Source/LodeNaVode/Controllers/PripravaController.cs
@@ -1,7 +1,6 @@
 ﻿using LodeNaVode.Models;
 using Microsoft.AspNetCore.Mvc;
 using System.Diagnostics;
-using LodeNaVode.Lode;
 
 namespace LodeNaVode.Controllers
 {
