+++ Source/LodeNaVode/Lode/tvary.txt
@@ -1,0 +1,18 @@
+A,S : -1,0;1,0;0,1
+B,S : -1,1;0,1;1,1;-1,-1;0,-1;1,-1
+C,S : 0,2;0,1;1,1;0,-1;1,-1;0,-2
+D,S : 0,2;0,1;0,-1;0,-2
+E,S : 0,1;1,0;0,-1;-1,0
+F,S : 0,2;0,1;1,0;2,0
+G,S : 1,2;0,2;1,1;0,1;0,-1;0,-2
+H,S : 0,1;-1,-1;0,-1;1,-1
+I,S : 1,2;0,2;0,1;1,0;1,-1;1,-2;0,-2
+J,S : -1,1;-1,0;1,0;1,-1
+K,M : 1,0;0,1
+L,M : 0,1;0,-1
+M,M :
+N,V : 0,1;0,-1;-1,0;-1,-1;-1,1;1,1;1,0;1,-1;-1,-2;1,-2;-1,2;1,2
+O,V : -1,2;-1,1;0,1;1,1;1,2;-1,-2;-1,-1;0,-1;1,-1;1,-2
+P,V : 1,1;0,1;2,1;-1,1;-2,1;0,-1;1,-1;2,-1;-1,-1;-2,-1
+Q,V : 1,-1;0,-1;-1,-1;0,-1;-1,1;0,1;1,1;0,2;1,2;0,-2;1,-2
+R,V : 0,-1;0,1;-1,1;1,1;-1,-1;1,-1;-1,2;1,2;1,-2;-1,-2
