+++ Source/LodeNaVode/Views/Tah/NalepkySiVyberTyMagor.cshtml
@@ -1,5 +1,5 @@
 ﻿@{
-    List<string> nalepky = LodeNaVode.Controllers.Engin.engine.GetLog.Data.nalepky;
+    List<string> nalepky = ViewBag.engine.GetLog.Data.nalepky;
 }
 <h1>vyber si</h1>
 <div style="text-align:left; font-family:monospace;">
