+++ Source/LodeNaVode/Views/Tah/Policko.cshtml
@@ -9,7 +9,7 @@
     string[,] config = Model.Item2;
 
     bool zobrazOvladani = LodeNaVode.Controllers.Pamet.oznacenaLod;
-    List<string> log = LodeNaVode.Controllers.Engin.engine.GetLog.GetEntireLog();
+    List<string> log = ViewBag.engine.GetLog.GetEntireLog();
     int logInd;
     if (log.Count() >= 5)
         logInd = log.Count() - 5;
